backend/requirements.txt
pypdf==5.1.0
pdfplumber==0.11.4
pytesseract==0.3.13
pdf2image==1.17.0
Pillow==11.0.0
python-docx==1.1.2
Flask==3.1.0
mammoth==1.8.0
beautifulsoup4==4.12.3
